+++ docs/changes.rst
@@ -86,6 +86,8 @@ Released: 2022-06-26
 * Added a dump_hmc_definition.py example script that dumps the resources of
   an HMC to a HMC definition file for use as a mock definition.
 
+* Improved mock support for password rules by creating default properties.
+
 
 Version 1.3.0
 ^^^^^^^^^^^^^